runtime-rs: Separate init_config() from new() for struct VsockDevice
As a follow-up for #8516, guest_cid and vhost_fd are not necessarily instantiated
via new(). Instead, the fields could be initialised later when they are really
used to construct QEMU parameters.
This commit is to separate init_config() from new() to initialise guest_cid
and vhost_fd and leave only the assignment of id for the existing function.

Signed-off-by: Hyounggyu Choi <Hyounggyu.Choi@ibm.com>