runtime-rs: Relocate const VM_ROOTFS_FILESYSTEM_* to kata-types
It is advisable to use a single rootfs type where it is used for a default
file system, rather than to have different const variables referring to the
same value for each hypervisor.

This is to relocate VM_ROOTFS_FILESYSTEM_* const variables to kata-types
and make them used for where a corresponding filesystem type is used.

Signed-off-by: Hyounggyu Choi <Hyounggyu.Choi@ibm.com>