packaging: Make TEE_PLATFORM configurable for guest-components
By default, all verifiers are compiled when TEE_PLATFORM is not set.
This is problematic for s390x, as the platform does not support the
required libraries for other verifiers (e.g., TDX, SGX, etc.).

The environment variable should be configurable, allowing the builder
to compile only the se-verifier for the platform using `TEE_PLATFORM=se`.

Signed-off-by: Hyounggyu Choi <Hyounggyu.Choi@ibm.com>