tests: Create key.bin according to directory structure
The current KBS deployment creates a file `key.bin` assuming that
`kustomization.yaml` is located in `overlays/`, which is valid if
no other platforms are enabled.

However, this does not hold true when the kustomize config is enabled
for multiple architectures. In such cases, the configuration file
should be located in `overlays/$(uname -m)`.
This commit addresses both scenarios to prevent operation failures
due to directory restructuring.

Signed-off-by: Hyounggyu Choi <Hyounggyu.Choi@ibm.com>