CI: Drop ci/install_rust.sh
As part of archiving the tests repo, we are eliminating the dependency on
`clone_tests_repo()`. One of the scripts using the function is `ci/install_rust.sh`.
It can be replaced with `tests/install_rust.sh`.

This commit removes the file and updates a path to the script accordingly.

Signed-off-by: Hyounggyu Choi <Hyounggyu.Choi@ibm.com>